+++ internal/agent/probes/snmptrap/receiver_security_test.go
@@ -159,7 +159,7 @@ func TestCollect_EmitsReceiverSelfMetrics(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Collect: %v", err)
 	}
-	got := map[string]float64{}
+	got := map[string]float32{}
 	for _, dp := range points {
 		got[dp.Name] = dp.Value
 	}
